Refactor metrics classes into separate module
